[3/3] base: Disable ADB over network on disconnect
ADB over network is not disabled before the next reboot, which creates
a potential security threat when this setting is left enabled and the
device gets re-connected to public wifi networks.

To mitigate that risk, introduce a new switch to disable ADB over network
as soon as the device gets disconnected from wifi.

Change-Id: Icc4eb748422b2f21cf5ae4be5054ca1c6c4e3d4e